Fix OS X 10.8-only behavior in the XGL plug-in. Remove some commented-out code on Cocoa code. Change some ints passed in Cocoa code to NSIntegers. Replace a deprecated function call to NSRunAlertPanelRelativeToWindow. Other minor fixes and improvements.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@86808 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/ConfigurationController.h
@@ -48,8 +48,8 @@ extern NSString *const memCardChangeNumberKey;
 - (IBAction)setVideoType:(id)sender;
 
 - (NSString *)keyForSender:(id)sender;
-+ (void)setMemoryCard:(int)theCard toPath:(NSString *)theFile;
-+ (void)setMemoryCard:(int)theCard toURL:(NSURL *)theURL;
++ (void)setMemoryCard:(NSInteger)theCard toPath:(NSString *)theFile;
++ (void)setMemoryCard:(NSInteger)theCard toURL:(NSURL *)theURL;
 
 + (void)mcdNewClicked:(id)sender;
 + (void)mcdChangeClicked:(id)sender;
